backend/pipeline/requirements.txt
flask
flask-cors
python-dotenv
requests
boto3
google-genai
